fix start job button
moved the logic and rendering of the JobBox into its own componet which fixed the Job start buttons
